Re-use the local file descriptor for a given batch
To avoid the overhead incurred of one (open, write, close) of the local
file per record, open the local file once per batch and store it in the
context for re-use across the batch.

I haven't finished processing the results of the pre-patch run against
the same data (because it's taking so long), but the run with this patch
took only 20 seconds:
```
poetry run target-s3-jsonl --config .config --input braze-output.singer  20.03s user 1.18s system 42% cpu 49.574 total
```

And, anecdotally, I was able to get a full meltano run of tap-braze ->
target-s3-jsonl run in ~11 minutes on my local machine.